requirements.txt
opencv-python
captacity
openai
pydub
numpy
azure-core